dev/combat_analyzer/requirements.txt
openai
python-dotenv
psycopg2-binary
